TRaits hobbies and hopes
